+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2021-08-22
+- Initial major release.
 - Added LogException methods.
 - Added DisableToggleKey method.
 - Changed the size of the input field so that the text is larger.